+++ src/Nethermind/Nethermind.Blockchain.Test/Validators/TxValidatorTests.cs
@@ -567,7 +567,7 @@ public class TxValidatorTests
     public void IsWellFormed_TransactionWithGasLimitExceedingEip7825Cap_ReturnsFalse()
     {
         Transaction tx = Build.A.Transaction
-            .WithGasLimit(Eip7825Constants.DefaultTxGasLimitCap + 1)
+            .WithGasLimit((ulong)(Eip7825Constants.DefaultTxGasLimitCap + 1))
             .WithChainId(TestBlockchainIds.ChainId)
             .SignedAndResolved().TestObject;
 
@@ -579,7 +579,7 @@ public class TxValidatorTests
         using (Assert.EnterMultipleScope())
         {
             Assert.That(result.AsBool, Is.False);
-            Assert.That(result.Error, Is.EqualTo(TxErrorMessages.TxGasLimitCapExceeded(tx.GasLimit, Eip7825Constants.DefaultTxGasLimitCap)));
+            Assert.That(result.Error, Is.EqualTo(TxErrorMessages.TxGasLimitCapExceeded((long)tx.GasLimit, Eip7825Constants.DefaultTxGasLimitCap)));
         }
     }
